fix: Deprecate Gazenot::new in favor of into_the_abyss
`Gazenot::new` is clearely inferior. I believe it should be deprecated
in version 0.1.1 and fully removed by version 0.2.0, maybe 0.3.0 if you
want to give consumers of this library enough time to prepare. Although
I don't belive anyone using the inferior `Gazenot::new` deserves such
priviledge. Still, we should warn them by first deprecating it and
adding an alias to it inside the documentation.

This is a joke commit, I hope you had a laugh and a wondeful day <3.

And of course, thank you to
[cargo-mommy](https://github.com/Gankra/cargo-mommy) for being very
encouraging and helping me submit this PR ( ˘ ³˘)❤  mommy is proud ~❤